+++ package.json
@@ -57,9 +57,7 @@
     "typescript": "^4.6.3"
   },
   "dependencies": {
-    "axios": "^0.27.2",
     "cross-fetch": "^4.1.0",
-    "json-bigint": "^1.0.0",
-    "web-streams-polyfill": "^4.1.0"
+    "json-bigint": "^1.0.0"
   }
 }
