+++ src/Record.ts
@@ -1,7 +1,7 @@
 import Stream, { Readable } from "stream";
 import { Buffer } from "buffer";
 import { WriteOptions } from "./Bucket";
-import { FetchClient } from "./http/HttpFetchClient";
+import { HttpClient } from "./http/HttpClient";
 
 export type LabelMap = Record<string, string | number | boolean | bigint>;
 
@@ -86,7 +86,7 @@ export class ReadableRecord {
 export class WritableRecord {
   private readonly bucketName: string;
   private readonly entryName: string;
-  private readonly fetchClient: FetchClient;
+  private readonly httpClient: HttpClient;
   private readonly options: WriteOptions;
 
   /**
@@ -97,11 +97,11 @@ export class WritableRecord {
     bucketName: string,
     entryName: string,
     options: WriteOptions,
-    fetchClient: FetchClient,
+    httpClient: HttpClient,
   ) {
     this.bucketName = bucketName;
     this.entryName = entryName;
-    this.fetchClient = fetchClient;
+    this.httpClient = httpClient;
     this.options = options;
   }
 
@@ -134,7 +134,7 @@ export class WritableRecord {
       throw new Error("Timestamp must be set");
     }
 
-    await this.fetchClient.post(
+    await this.httpClient.post(
       `/b/${bucketName}/${entryName}?ts=${options.ts}`,
       data,
       headers,
