--- src/http/HttpFetchClient.ts
@@ -1,172 +1,0 @@
-import JSONbig from "json-bigint";
-import fetch from "cross-fetch";
-import { Readable } from "stream";
-import { Agent as HttpsAgent } from "https";
-import { ClientOptions } from "../Client";
-import { APIError } from "../APIError";
-import { isBrowser } from "../utils/env";
-
-const bigJson = JSONbig({ alwaysParseAsBig: false, useNativeBigInt: true });
-
-export type ValidResponse = object | string | ReadableStream<Uint8Array>;
-
-export type FetchResult<T extends ValidResponse = ValidResponse> = {
-  data: T;
-  headers: Headers;
-  status: number;
-};
-
-export class FetchClient {
-  private baseURL: string;
-  private timeout?: number;
-  private headers: HeadersInit;
-  private agent?: HttpsAgent;
-
-  constructor(url: string, options: ClientOptions = {}) {
-    this.baseURL = `${url}/api/v1`;
-    this.timeout = options.timeout;
-    this.headers = { Authorization: `Bearer ${options.apiToken}` };
-
-    if (!isBrowser && options.verifySSL === false) {
-      this.agent = new HttpsAgent({ rejectUnauthorized: false });
-    }
-  }
-
-  // ---------- request overloads ----------
-
-  private async request(
-    method: "HEAD",
-    url: string,
-    body?: unknown,
-    headers?: HeadersInit,
-  ): Promise<FetchResult<Record<string, never>>>;
-
-  private async request<T extends ValidResponse>(
-    method: "GET",
-    url: string,
-    body?: unknown,
-    headers?: HeadersInit,
-  ): Promise<FetchResult<T>>;
-
-  private async request<T extends ValidResponse>(
-    method: string,
-    url: string,
-    body?: unknown,
-    headers?: HeadersInit,
-  ): Promise<FetchResult<T>>;
-
-  // ---------- request implementation ----------
-  private async request<T extends ValidResponse>(
-    method: string,
-    url: string,
-    body?: unknown,
-    headers?: HeadersInit,
-  ): Promise<FetchResult<T>> {
-    const controller = new AbortController();
-
-    const { timeout } = this;
-    if (timeout) setTimeout(() => controller.abort(), timeout);
-
-    const init: RequestInit = {
-      method,
-      headers: { ...this.headers, ...headers },
-      body: this.encodeBody(body),
-      signal: controller.signal,
-      // @ts-ignore
-      agent: this.agent, // Node.js specific, fine to ts-ignore
-    };
-
-    const response = await fetch(`${this.baseURL}${url}`, init).catch((err) => {
-      throw new APIError(err.message, undefined, err);
-    });
-
-    if (!response.ok) {
-      const message =
-        response.headers.get("x-reduct-error") || response.statusText;
-      throw new APIError(message, response.status, { response });
-    }
-
-    const data = (await this.parseResponse(response)) as T;
-
-    return {
-      data,
-      headers: response.headers,
-      status: response.status,
-    };
-  }
-
-  private encodeBody(data?: unknown): BodyInit | undefined {
-    if (
-      data === undefined ||
-      typeof data === "string" ||
-      Buffer.isBuffer(data) ||
-      data instanceof Uint8Array ||
-      data instanceof Readable ||
-      data instanceof ArrayBuffer ||
-      data instanceof Blob
-    ) {
-      return data as BodyInit;
-    }
-    return bigJson.stringify(data);
-  }
-
-  private async parseResponse(
-    res: Response,
-  ): Promise<object | string | ReadableStream<Uint8Array>> {
-    if (res.status === 204) return {};
-    const ct = res.headers.get("content-type") ?? "";
-
-    if (!res.body) return {};
-
-    if (ct.startsWith("application/json")) {
-      const text = await res.text();
-      return text ? bigJson.parse(text) : {};
-    }
-    if (ct.startsWith("text/")) {
-      return res.text();
-    }
-    return res.body;
-  }
-
-  // ---------- helpers ----------
-  get<T extends ValidResponse = ValidResponse>(
-    url: string,
-  ): Promise<FetchResult<T>> {
-    return this.request<T>("GET", url);
-  }
-
-  post<T extends ValidResponse = ValidResponse>(
-    url: string,
-    data?: unknown,
-    headers?: HeadersInit,
-  ): Promise<FetchResult<T>> {
-    return this.request<T>("POST", url, data, headers);
-  }
-
-  put<T extends ValidResponse = ValidResponse>(
-    url: string,
-    data?: unknown,
-    headers?: HeadersInit,
-  ): Promise<FetchResult<T>> {
-    return this.request<T>("PUT", url, data, headers);
-  }
-
-  patch<T extends ValidResponse = ValidResponse>(
-    url: string,
-    data?: unknown,
-    headers?: HeadersInit,
-  ): Promise<FetchResult<T>> {
-    return this.request<T>("PATCH", url, data, headers);
-  }
-
-  delete<T extends ValidResponse = ValidResponse>(
-    url: string,
-    headers?: HeadersInit,
-  ): Promise<FetchResult<T>> {
-    return this.request<T>("DELETE", url, undefined, headers);
-  }
-
-  head(url: string): Promise<FetchResult<Record<string, never>>> {
-    return this.request("HEAD", url);
-  }
-}
